Convert percent slope band to int16 and scale it.

+++ dswe/src/output.h
@@ -53,8 +53,8 @@ add_ps_band_product
     char *short_name,
     char *long_name,
     int min_range,
-    float max_range,
-    float *data
+    int max_range,
+    int16_t *data
 );
